+++ packages/core/src/components/viewer-grid/viewer-grid.interfaces.ts
@@ -7,10 +7,24 @@ export interface ViewerGridChangeSizeEvent {
   currentSize: MainSize;
 }
 
-export interface FilterpanelEvent {
-  originalEvent: MouseEvent;
+export interface ViewerGridCloseOverlayEvent {
+  originalEvent: MouseEvent | Event;
+}
+
+export interface ViewerGridFilterpanelCancelEvent {
+  originalEvent: MouseEvent | Event;
+}
+
+export interface ViewerGridFilterpanelApplyEvent {
+  originalEvent: MouseEvent | Event;
 }
 
 export type MainSize = "small" | "medium" | "large";
 
 export type LabelSizeMap = { [key in MainSize]: string };
+
+export const tabs = ["main", "map"] as const;
+
+export type Tabs = (typeof tabs)[number];
+
+export type TabLabelMap = { [key in Tabs]: string };