+++ packages/core/src/components/viewer-grid/viewer-grid.tsx
@@ -1,7 +1,19 @@
 import { h, Component, Prop, State, Host, Element, Event, EventEmitter, Watch } from "@stencil/core";
-import { FocusTrap, createFocusTrap } from "focus-trap";
-import { ViewerGridFilterpanelButtons } from "./viewer-grid-filterpanel-buttons";
-import { FilterpanelEvent, LabelSizeMap, MainSize, ViewerGridChangeSizeEvent } from "./viewer-grid.interfaces";
+import clsx from "clsx";
+
+import {
+  MainSize,
+  TabLabelMap,
+  Tabs,
+  ViewerGridChangeSizeEvent,
+  ViewerGridCloseOverlayEvent,
+  ViewerGridFilterpanelApplyEvent,
+  ViewerGridFilterpanelCancelEvent,
+  tabs,
+} from "./viewer-grid.interfaces";
+import { Filterpanel, MainPanel, Overlay } from "./components";
+
+const tabViewWidth = 768 + 40;
 
 @Component({
   tag: "dso-viewer-grid",
@@ -9,13 +21,14 @@ import { FilterpanelEvent, LabelSizeMap, MainSize, ViewerGridChangeSizeEvent } f
   shadow: true,
 })
 export class ViewerGrid {
-  private sizeLabelMap: LabelSizeMap = {
-    small: "smal",
-    medium: "middel",
-    large: "breed",
+  private mediaCondition = `(min-width: ${tabViewWidth}px)`;
+
+  private tabLabelMap: TabLabelMap = {
+    main: "Hoofdpaneel",
+    map: "Kaart",
   };
 
-  private mapPanel?: HTMLDivElement;
+  private mainPanel?: HTMLDivElement;
 
   /**
    * Set to true when filterpanel should show.
@@ -40,23 +53,29 @@ export class ViewerGrid {
   @State()
   mainSize: MainSize = "large";
 
+  @State()
+  tabView = window.innerWidth < tabViewWidth;
+
+  @State()
+  activeTab: Tabs = "main";
+
   /**
    * Emitted when user wants to close the overlay.
    */
   @Event()
-  dsoCloseOverlay!: EventEmitter<MouseEvent | KeyboardEvent>;
+  dsoCloseOverlay!: EventEmitter<ViewerGridCloseOverlayEvent>;
 
   /**
    * Emitted when user cancels filterpanel.
    */
   @Event()
-  dsoFilterpanelCancel!: EventEmitter<FilterpanelEvent>;
+  dsoFilterpanelCancel!: EventEmitter<ViewerGridFilterpanelCancelEvent>;
 
   /**
    * Emitted when user applies filterpanel options.
    */
   @Event()
-  dsoFilterpanelApply!: EventEmitter<FilterpanelEvent>;
+  dsoFilterpanelApply!: EventEmitter<ViewerGridFilterpanelApplyEvent>;
 
   /**
    * Emitted before and after main size animation. Inspect `detail` property for more information.
@@ -67,17 +86,13 @@ export class ViewerGrid {
   @Element()
   host!: HTMLDsoViewerGridElement;
 
-  private filterpanel: HTMLElement | undefined;
+  private filterpanel: HTMLDialogElement | undefined;
 
   private filterpanelSlot: HTMLElement | null = null;
 
-  private filterpanelFocustrap: FocusTrap | undefined;
-
-  private overlay: HTMLDivElement | undefined;
+  private overlay: HTMLDialogElement | undefined;
 
   private overlaySlot: HTMLDivElement | null = null;
-
-  private overlayFocustrap: FocusTrap | undefined;
 
   @Watch("mainSize")
   mainSizeWatcher(currentSize: MainSize, previousSize: MainSize) {
@@ -87,7 +102,7 @@ export class ViewerGrid {
       currentSize,
     });
 
-    this.mapPanel?.addEventListener(
+    this.mainPanel?.addEventListener(
       "transitionend",
       (e) => {
         if (e.propertyName === "flex-basis") {
@@ -102,6 +117,32 @@ export class ViewerGrid {
     );
   }
 
+  @Watch("filterpanelOpen")
+  filterpanelOpenWatcher(open: boolean) {
+    if (!this.filterpanelSlot) {
+      console.warn("slot 'filterpanel' has not been set");
+    }
+
+    if (open) {
+      this.filterpanel?.showModal();
+    } else {
+      this.filterpanel?.close();
+    }
+  }
+
+  @Watch("overlayOpen")
+  overlayOpenWatcher(open: boolean) {
+    if (!this.overlaySlot) {
+      console.warn("slot 'overlay' has not been set");
+    }
+
+    if (open) {
+      this.overlay?.showModal();
+    } else {
+      this.overlay?.close();
+    }
+  }
+
   private shrinkMain = () => {
     this.mainSize = this.mainSize === "large" ? "medium" : "small";
   };
@@ -110,68 +151,24 @@ export class ViewerGrid {
     this.mainSize = this.mainSize === "small" ? "medium" : "large";
   };
 
-  private updateFocusTrap() {
-    if (this.filterpanelOpen && this.overlayOpen) {
-      return;
-    }
-
-    if (this.filterpanelFocustrap) {
-      if (this.filterpanelOpen && !this.filterpanel?.hidden) {
-        this.filterpanelFocustrap.activate();
-        this.host.addEventListener("keydown", this.keyDownListener);
-      } else {
-        this.filterpanelFocustrap.deactivate();
-        this.host.removeEventListener("keydown", this.keyDownListener);
-      }
-    }
-
-    if (this.overlayFocustrap) {
-      if (this.overlayOpen && !this.overlay?.hidden) {
-        this.overlayFocustrap.activate();
-        this.host.addEventListener("keydown", this.keyDownListener);
-      } else {
-        this.overlayFocustrap.deactivate();
-        this.host.removeEventListener("keydown", this.keyDownListener);
-      }
-    }
-  }
-
-  private keyDownListener = (event: KeyboardEvent) => {
-    if (event.key !== "Escape") {
-      return;
-    }
-
-    this.dsoCloseOverlay.emit(event);
-  };
+  private changeListener = (largeScreen: MediaQueryListEvent) => (this.tabView = !largeScreen.matches);
 
   connectedCallback() {
+    window.matchMedia(this.mediaCondition).addEventListener("change", this.changeListener);
+
     this.filterpanelSlot = this.host.querySelector<HTMLDivElement>("div[slot='filterpanel']");
 
     this.overlaySlot = this.host.querySelector<HTMLDivElement>("div[slot='overlay']");
   }
 
   componentDidLoad() {
-    if (this.filterpanel && this.filterpanelSlot) {
-      this.filterpanelFocustrap = createFocusTrap([this.filterpanel, this.filterpanelSlot], {
-        escapeDeactivates: false,
-        allowOutsideClick: true,
-        tabbableOptions: {
-          getShadowRoot: true,
-        },
-      });
+    if (this.filterpanelOpen && this.filterpanelSlot) {
+      this.filterpanel?.showModal();
     }
 
-    if (this.overlay && this.overlaySlot) {
-      this.overlayFocustrap = createFocusTrap([this.overlay, this.overlaySlot], {
-        escapeDeactivates: false,
-        allowOutsideClick: true,
-        tabbableOptions: {
-          getShadowRoot: true,
-        },
-      });
+    if (this.overlayOpen && this.overlaySlot) {
+      this.overlay?.showModal();
     }
-
-    this.updateFocusTrap();
   }
 
   componentWillLoad() {
@@ -180,83 +177,57 @@ export class ViewerGrid {
     }
   }
 
-  componentDidUpdate() {
-    this.updateFocusTrap();
-  }
-
   disconnectedCallback() {
-    this.overlayFocustrap?.deactivate();
-    this.filterpanelFocustrap?.deactivate();
-
-    this.host.removeEventListener("keydown", this.keyDownListener);
+    window.matchMedia(this.mediaCondition).removeEventListener("change", this.changeListener);
   }
 
-  private handleFilterpanelApply(mouseEvent: MouseEvent) {
+  private handleFilterpanelApply(mouseEvent: MouseEvent | Event) {
     this.dsoFilterpanelApply.emit({ originalEvent: mouseEvent });
   }
 
-  private handleFilterpanelCancel(mouseEvent: MouseEvent) {
+  private handleFilterpanelCancel(mouseEvent: MouseEvent | Event) {
     this.dsoFilterpanelCancel.emit({ originalEvent: mouseEvent });
   }
 
   render() {
     return (
       <Host {...{ [this.mainSize]: true }}>
-        <div class="dso-map-panel" ref={(element) => (this.mapPanel = element)}>
-          <div class="sizing-buttons">
-            <span class="sr-only" aria-live="polite" aria-atomic="true">
-              Breedte tekstpaneel: {this.sizeLabelMap[this.mainSize]}
-            </span>
-            <button type="button" class="shrink" disabled={this.mainSize === "small"} onClick={this.shrinkMain}>
-              <span class="sr-only">Tekstpaneel smaller maken</span>
-              <dso-icon icon="chevron-left"></dso-icon>
-            </button>
-            <button type="button" class="expand" disabled={this.mainSize === "large"} onClick={this.expandMain}>
-              <span class="sr-only">Tekstpaneel breder maken</span>
-              <dso-icon icon="chevron-right"></dso-icon>
-            </button>
-          </div>
-          <div class="main">
-            <slot name="main" />
-          </div>
-        </div>
-        <div
-          id="filterpanel"
-          class="filterpanel"
-          hidden={!this.filterpanelOpen || !this.filterpanelSlot}
+        {this.tabView && (
+          <nav class="dso-navbar">
+            <ul class="dso-nav dso-nav-sub">
+              {tabs.map((tab) => (
+                <li key={tab} class={clsx({ "dso-active": this.activeTab === tab })}>
+                  <button type="button" class="dso-tertiary" onClick={() => (this.activeTab = tab)}>
+                    {this.tabLabelMap[tab]}
+                  </button>
+                </li>
+              ))}
+            </ul>
+          </nav>
+        )}
+        {(!this.tabView || (this.tabView && this.activeTab === "main")) && (
+          <MainPanel
+            ref={(element: HTMLDivElement | undefined) => (this.mainPanel = element)}
+            tabView={this.tabView}
+            mainSize={this.mainSize}
+            shrinkMain={this.shrinkMain}
+            expandMain={this.expandMain}
+          ></MainPanel>
+        )}
+        <Filterpanel
           ref={(element) => (this.filterpanel = element)}
-        >
-          <h1>Uw keuzes</h1>
-          <ViewerGridFilterpanelButtons
-            onApply={(e) => this.handleFilterpanelApply(e)}
-            onCancel={(e) => this.handleFilterpanelCancel(e)}
-          />
-          <slot name="filterpanel" />
-          <ViewerGridFilterpanelButtons
-            onApply={(e) => this.handleFilterpanelApply(e)}
-            onCancel={(e) => this.handleFilterpanelCancel(e)}
-          />
-        </div>
-        <div class="map">
-          <slot name="map" />
-        </div>
-        <div hidden={!this.overlayOpen || !this.overlaySlot} class="dimscreen"></div>
-        <div
-          class="overlay"
-          hidden={!this.overlayOpen || !this.overlaySlot}
+          onApply={(e) => this.handleFilterpanelApply(e)}
+          onCancel={(e) => this.handleFilterpanelCancel(e)}
+        ></Filterpanel>
+        {(!this.tabView || (this.tabView && this.activeTab === "map")) && (
+          <div class="map">
+            <slot name="map" />
+          </div>
+        )}
+        <Overlay
           ref={(element) => (this.overlay = element)}
-        >
-          <button type="button" class="overlay-close-button" onClick={(e) => this.dsoCloseOverlay.emit(e)}>
-            <dso-icon icon="times"></dso-icon>
-            <span class="sr-only">sluiten</span>
-          </button>
-          <slot name="overlay" />
-          {/* This button is needed for the `focus-trap` library to function correctly. It is never focused. */}
-          <button aria-hidden="true" type="button" class="overlay-close-button" style={{ zIndex: "-100" }}>
-            <dso-icon icon="times"></dso-icon>
-            <span class="sr-only">sluiten</span>
-          </button>
-        </div>
+          dsoCloseOverlay={(e) => this.dsoCloseOverlay.emit({ originalEvent: e })}
+        ></Overlay>
       </Host>
     );
   }